+++ src/transposition.c
@@ -169,7 +169,7 @@ PKEntry* getPKEntry(PKTable *pktable, uint64_t pkhash) {
     return pkentry->pkhash == pkhash ? pkentry : NULL;
 }
 
-void storePKEntry(PKTable *pktable, uint64_t pkhash, uint64_t passed, int eval) {
+void storePKEntry(PKTable *pktable, uint64_t pkhash, uint64_t passed, int64_t eval) {
     PKEntry *pkentry = &pktable->entries[pkhash >> PKT_HASH_SHIFT];
     pkentry->pkhash = pkhash;
     pkentry->passed = passed;
