+++ src/transposition.h
@@ -62,7 +62,7 @@ struct TTable {
 struct PKEntry {
     uint64_t pkhash;
     uint64_t passed;
-    uint64_t eval;
+    int64_t eval;
 };
 
 struct PKTable {
@@ -79,4 +79,4 @@ int getTTEntry(uint64_t hash, uint16_t *move, int *value, int *eval, int *depth,
 void storeTTEntry(uint64_t hash, uint16_t move, int value, int eval, int depth, int bound);
 
 PKEntry* getPKEntry(PKTable *pktable, uint64_t pkhash);
-void storePKEntry(PKTable *pktable, uint64_t pkhash, uint64_t passed, int eval);
+void storePKEntry(PKTable *pktable, uint64_t pkhash, uint64_t passed, int64_t eval);
